requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
loguru==0.7.0
Pillow==10.0.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
